清除warning
git-svn-id: http://192.168.20.11/sequoiadb/trunk@21079 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/bson/bson-inl.h
@@ -731,7 +731,7 @@ namespace bson {
             else
             {
                s << "{ \"$date\": "  ;
-               sprintf ( buffer, "%lld", timer * 1000 ) ;
+               sprintf ( buffer, "%lld", (unsigned long long)(timer * 1000) ) ;
                s << buffer << " }" ;
             }
             break ;
